// packages/spark-bff/src/generated/typings/accessServiceV2.ts
export interface definitions {
  AccessResponse: definitions['Response'] & {
    lines: definitions['SingleAccessRecord'][];
    accounts: definitions['SingleAccessRecord'][];
    identities: definitions['SingleAccessRecord'][];
  };
  AccessPermissionResponse: definitions['Response'] & {
    access: definitions['SinglePermissionRecord'][];
  };
  SingleAccessRecord: {
    /** The line of the access record (only present in line-typed records) */
    lineNumber?: string;
    /** The customer number of the access record (present in account & line-typed records) */
    customerNumber?: string;
    /** The account of the access record (present in account & line-typed records; PP prefixed account numbers should only show up in line-typed records) */
    accountNumber?: string;
    /** Revoked is only for accounts, and will only show if such accounts are requested.  NotFound is only for accounts & lines, ditto. */
    status: 'Active' | 'Revoked' | 'NotFound';
    /** The type of the access record */
    type: 'Line' | 'Account' | 'Identity';
    /** The type of the line (if this is a line level access record), or of all the lines under the account (if this is an account level access record).  Strictly speaking, this is an account-level attribute, but we mostly use it at line level, so I have phrased it that way. */
    balanceManagement?: 'PREPAID' | 'POSTPAID';
    /** The role of the access record */
    role: string;
    /** The user-friendly alias for the access record */
    aliasName?: string;
    /** The line numbers associated with this account (will only be present in non-revoked account-typed records).  If this is returned via a line-level access request, this will contain at most the requested line number. */
    lineNumbers?: string[];
    /** The list of rights (aka eligibilities) that the role grants (will only be present in account & line records). */
    rights?: string[];
  };
// ...
  Response: {
    /** List of Messages */
    messages: definitions['Message'][];
  };
  Message: {
    message: string;
    code: number;
  };
}